Bugfix: make recipe run
Had a buggy case/when statement - not sure if this broke in a new
ruby version or if this never worked.

Also removed some apparently outdated configuration settings from
the erb template.

Ensured the config directory is created, before we try to add a
config file into it.

Refactored some recipe code.