Adds DATABASE_URL to app's environment
The ``DATABASE_URL`` is a environment variable used e.g. in the Ruby ecosystem. It is a composite of all DB details in form of a URL like this:

```
mysql2://user:password@host:port/database?reconnect=true
```

This patch creates that variable in case that all necessary parts are set in the DB config.